simpler related entities link
