+++ datahub-web-react/src/app/entity/shared/ExternalUrlButton.tsx
@@ -1,27 +1,10 @@
-import { ArrowRightOutlined } from '@ant-design/icons';
-import { Button } from 'antd';
 import React from 'react';
-import styled from 'styled-components/macro';
 import { EntityType } from '../../../types.generated';
 import analytics, { EventType, EntityActionType } from '../../analytics';
+import UrlButton from './UrlButton';
 
 const GITHUB_LINK = 'github.com';
 const GITHUB = 'GitHub';
-
-const ExternalUrlWrapper = styled.span`
-    font-size: 12px;
-`;
-
-const StyledButton = styled(Button)`
-    > :hover {
-        text-decoration: underline;
-    }
-    &&& {
-        padding-bottom: 0px;
-    }
-    padding-left: 12px;
-    padding-right: 12px;
-`;
 
 interface Props {
     externalUrl: string;
@@ -46,17 +29,8 @@ export default function ExternalUrlButton({ externalUrl, platformName, entityTyp
     }
 
     return (
-        <ExternalUrlWrapper>
-            <StyledButton
-                type="link"
-                href={externalUrl}
-                target="_blank"
-                rel="noreferrer noopener"
-                onClick={sendAnalytics}
-            >
-                {displayedName ? `View in ${displayedName}` : 'View link'}{' '}
-                <ArrowRightOutlined style={{ fontSize: 12 }} />
-            </StyledButton>
-        </ExternalUrlWrapper>
+        <UrlButton href={externalUrl} onClick={sendAnalytics}>
+            {displayedName ? `View in ${displayedName}` : 'View link'}
+        </UrlButton>
     );
 }
